♻️ 컴포넌트 분리 및 리팩토링
