+++ src/app/page.tsx
@@ -6,7 +6,7 @@ import { TAGLIST } from "@/constants/Tag";
 import { AnimatePresence } from "framer-motion";
 import { useEffect } from "react";
 import usePageStore from "../stores/usePageStore";
-import usePageAnimateRouter from "@/hooks/usePageAnimateRouter";
+import usePageAnimateRouter from "@/hooks/common/usePageAnimateRouter";
 import Motion from "@/components/motion/Motion";
 
 export default function Home() {
